トップの広告が押せるように修正 レビューポップアップを表示

+++ astronauts/Classes/GameScene.m
@@ -12,7 +12,6 @@
 #import "PhysicsLayer.h"
 #import "ResultLayer.h"
 #import "BackgroundLayer.h"
-#import "AdLayer.h"
 #import "OALSimpleAudio.h"
 #import <GameKit/GameKit.h>
 
@@ -196,7 +195,7 @@ static GameScene *_scene = nil;
                     }
                     int highScore;
                     if (scores != nil) {
-                        highScore = ((GKScore *)[scores objectAtIndex:0]).value;
+                        highScore = (int)((GKScore *)[scores objectAtIndex:0]).value;
                     } else {
                         // 初回プレイ時はscoresがnilなので0ハイスコアを0とする。
                         highScore = 0;
